backend/pyproject.toml
[tool.poetry]
name = "integraflow"
version = "0.0.1"
description = "Integraflow provides tools to redefine customer experience with organic feedback and behavioural data in real-time."
authors = [ "Integraflow <dev@useintegraflow.io>" ]
license = "MIT"
readme = "README.md"
homepage = "https://useintegraflow.com/"
repository = "https://github.com/Integraflow/integraflow"
documentation = "https://docs.useintegraflow.io/"

  [tool.poetry.dependencies]
  python = "~3.9"
  babel = ">=2.8,<2.14"
  boto3 = "^1.28"
  dj-database-url = "^2"
  dj-email-url = "^1"
  django = {version = "^3.2.22", extras = ["bcrypt"]}
  django-countries = "^7.2"
  django-filter = "^23.1"
  django-measurement = "^3.0"
  django-mptt = "^0"
  django-phonenumber-field = ">=4,<8"
  django-prices = "^2.3"
  django-stubs-ext = "^4.2.1"
  draftjs-sanitizer = "^1.0.0"
  faker = ">=4.1,<20.0"
  freezegun = "^1"
  google-cloud-storage = "^2.0.0"
  google-i18n-address = "^3.1.0"
  html-to-draftjs = "^1.0.1"
  markdown = "^3.1.1"
  petl = "1.7.14"
  opentracing = "^2.3.0"
  phonenumberslite = "^8.12.25"
  prices = "^1.0"
  psycopg2 = "^2.8.3"
  requests = "^2.22"
  sentry-sdk = "^1.28"
  text-unidecode = "^1.2"
  # libpango >=1.44 is required
  weasyprint = ">=53.0"
  oauthlib = "^3.1"
  jaeger-client = "^4.5.0"
  openpyxl = "^3.0.3"
  django-cache-url = "^3.1.2"
  pyjwt = "2.5.0"
  python-json-logger = ">=0.1.11,<2.1.0"
  pytimeparse = "^1.1.8"
  python-magic = "^0.4.27"
  django-redis = "^5.0.0"
  google-cloud-pubsub = ">=1.7,<3.0"
  gunicorn = "^21.2.0"
  pybars3 = "^0.9.7"
  html2text = "^2020.1.16"
  sendgrid = "^6.7.1"
  micawber = "^0.5.2"
  django-celery-beat = "^2.2.1"
  posuto = "^2023.2.0"
  cryptography = "^41.0.4"
  graphene = "<3.0"
  uvicorn = {extras = ["standard"], version = "^0.23.1"}
  Authlib = "^1.0.0"
  pillow-avif-plugin = "^1.3.1"
  semantic-version = "^2.10.0"
  requests-hardened = "1.0.0b2"
  google-auth-oauthlib = "^1.1.0"
  django-cors-headers = "^4.3.0"

    [tool.poetry.dependencies.celery]
    version = ">=4.4.5,<6.0.0"
    extras = [ "redis" ]

    [tool.poetry.dependencies.django-storages]
    version = "^1.11"
    extras = [ "google" ]

    [tool.poetry.dependencies.python-magic-bin]
    version = "^0.4.14"
    platform = "win32"

  [tool.poetry.dev-dependencies]
  black = "23.9.1"
  coverage = "^7.2"
  django-debug-toolbar = "^4.0"
  django-debug-toolbar-request-history = "^0"
  django-graphiql-debug-toolbar = "^0.2.0"
  django-extensions = "^3.1.2"
  flake8 = "^4.0.1"
  isort = "^5.12.0"
  pre-commit = "^3.4"
  pycodestyle = "^2.5"
  pydocstyle = "^6.1"
  pylint = "^2.8.3"
  pylint-celery = "^0"
  pylint-django = "^2.4.4"
  pylint-plugin-utils = "^0"
  pytest = "^7.0.1"
  pytest-asyncio = "^0.21.0"
  pytest-celery = "^0.0.0"
  pytest-cov = "^4.0.0"
  pytest-django = "4.5.2"
  pytest-django-queries = "~1.2"
  pytest-mock = "^3.6.1"
  pytest-recording = "^0.13.0"
  pytest-socket = "^0.6.0"
  pytest-xdist = "^3.0.2"
  mypy = "1.4.1"
  mypy-extensions = "^1.0.0"
  pywatchman = "^1.4.1"
  toronado = "0.1.0"
  types-mock = "^5.0.0.5"
  types-redis = "^4.2.6"
  types-requests = "^2.31.0"
  types-pytz = "^2023.3.0"
  types-pkg-resources = "^0.1.3"
  types-python-dateutil = "^2.8.19"
  django-stubs = "1.16.0"
  before_after = "^1.0.1"
  types-certifi = "^2021.10.8"
  types-freezegun = "^1.1.7"
  types-six = "^1.16.17"
  vcrpy = ">=4.0,<5.0"
  fakeredis = "^2.10"
  semgrep = ">=1.34.0"

[tool.black]
target_version = [ "py35", "py36", "py37", "py38" ]
include = "\\.pyi?$"
exclude = """
/(\\.git/
  |\\.eggs
  |\\.hg
  |__pycache__
  |\\.cache
  |\\.ipynb_checkpoints
  |\\.mypy_cache
  |\\.pytest_cache
  |\\.tox
  |\\.venv
  |node_modules
  |_build
  |buck-out
  |build
  |dist
  |media
  |infrastructure
  |templates
  |locale
)/
"""
